Changed methods to return List in stead of Iterable

+++ playlist_generator/src/main/java/track_ninja/playlist_generator/repositories/GenreRepository.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Repository;
 import track_ninja.playlist_generator.models.Genre;
 
 @Repository
-public interface GenreRepository extends CrudRepository<Genre, Long> {
+public interface GenreRepository extends CrudRepository<Genre, Integer> {
     boolean existsByName(String name);
 
     Genre findByName(String genreName);
